+++ doc.go
@@ -1,6 +1,6 @@
 // Package archepixel provides OpenGL graphics and gonum plots for the Arche ECS using the Pixel game engine.
 //
 // See the sub-packages:
-//   - Live plotting drawers using gonum/plot -- [github.com/mlange-42/arche-pixel/plot]
-//   - OpenGL window system for free drawing -- [github.com/mlange-42/arche-pixel/window]
+//   - Live plotting and other visualizations -- [github.com/mlange-42/arche-pixel/plot]
+//   - OpenGL window system, for hosting visualizations -- [github.com/mlange-42/arche-pixel/window]
 package archepixel