+++ plot/doc.go
@@ -1,2 +1,5 @@
-// Package plot provides live plotting drawers for window.Window, using the gonum/plot package.
+// Package plot provides live plotting drawers for window.Window,
+// using the gonum/plot package or direct OpenGL drawing.
+//
+// For the window, see package [github.com/mlange-42/arche-pixel/window].
 package plot