+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * `Monitor` drawer for visualizing world and performance statistics (#10)
 * Windows are resizable (#10)
 * Window title can be set at construction time (#11)
-* Adds interface `InputHandler` for handling user input (#12)
+* Adds method `UpdateInputs` to `Drawer` interface, for handling user input (#12, #14)
 * Adds `Controls` plot and input handler for controlling simulation speed and pause via GUI or keyboard (#12)
 * `Image` and `ImageRGB` auto-scale when no explicit scale is given (#13)
 * Adds Method `window.Scale` to calculate scaling like in `Image` and `ImageRGB` (#13)